requirements-prod.txt
# Core Dependencies
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.3.0
django-filter==24.3
drf-yasg==1.21.7

# Production Server & Static Files
gunicorn==21.2.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Async Tasks & Caching
celery==5.4.0
redis==5.0.3

# Monitoring (Optional but recommended)
# sentry-sdk==1.40.6
